Created AuthComponent for use on all 3 panels of the landing page
