+++ client/src/pages/Landing/PanelOne.jsx
@@ -1,23 +1,12 @@
 import { Link } from 'react-router-dom';
-import { useState } from 'react';
 import Logo from '../../assets/tarot_logo.png';
 import MoonUp from '../../assets/Up_moon.png';
 import MoonDown from '../../assets/Down_moon.png';
 import BackgroundImage from '../../assets/TarotHeader1.png';
 import StarsBackground from '../Landing/Assets/Stars.png';
-import LoginModal from './Login_Signup/LoginModal';
-import SignupModal from './Login_Signup/SignUpModal';
+import AuthComponent from './Login_Signup/AuthComponent';
 
 const PanelOne = () => {
-    const [loginOpen, setLoginOpen] = useState(false);
-    const [signupOpen, setSignupOpen] = useState(false);
-
-    const handleLoginOpen = () => setLoginOpen(true);
-    const handleLoginClose = () => setLoginOpen(false);
-
-    const handleSignUpOpen = () => setSignupOpen(true);
-    const handleSignUpClose = () => setSignupOpen(false);
-
     return (
         <div
             className='panel-container'
@@ -44,18 +33,7 @@ const PanelOne = () => {
                 Embark on an Enlightening <br /> Journey with Tarot
             </h1>
             <img src={MoonDown} alt='Moon decorative element' />
-            <div className='button-container'>
-                <div>
-                    <button className='button' onClick={handleLoginOpen}>
-                        Login
-                    </button>
-                    <button className='button' onClick={handleSignUpOpen}>
-                        Sign Up
-                    </button>
-                </div>
-            </div>
-            <LoginModal open={loginOpen} handleClose={handleLoginClose} />
-            <SignupModal open={signupOpen} handleClose={handleSignUpClose} />
+            <AuthComponent />
         </div>
     );
 };
